Data Analysis/requirements.txt
matplotlib.pyplot
pandas
termcolor